feat: authentification and data access

+++ backend/repositories/todolist.repo.js
@@ -1,52 +1,53 @@
 const { pool, query } = require("../config/database");
+const { use } = require("../routes/todolist.route");
 
 class Method {
 
-    async create(data) {
+    async create(data, userId) {
         // créer une to-do list (titre, couleur, pinned)
         const { title, pinned, color } = data;
 
         const query = `INSERT INTO to_do_lists (user_id, title, pinned, color) VALUES ($1, $2, $3, $4) RETURNING *`;
 
-        const values = [1, title, pinned || false, color || 'grey'];
+        const values = [userId, title, pinned || false, color || 'grey'];
 
         const result = await pool.query(query, values);
 
         return result.rows[0];
     }
 
-    async readAll() {
+    async readAll(userId) {
         // lire toutes les to-do lists : pour la liste des to-do lists
-        const result = await query(`SELECT * FROM to_do_lists`);
+        const result = await query(`SELECT * FROM to_do_lists WHERE user_id = $1`, [userId]);
         return result.rows;
     }
 
-    async readOne(id) {
+    async readOne(id, userId) {
         // lire une seule to-do list : pour les cartes to-do list
-        const result = await query(`SELECT * FROM to_do_lists WHERE id = $1`, [id]);
+        const result = await query(`SELECT * FROM to_do_lists WHERE id = $1 AND user_id = $2`, [id, userId]);
         return result.rows;
     }
 
-    async update(id, data) {
+    async update(id, data, userId) {
         // pour mettre à jour une to-do list (son titre, sa couleur, pin ou pas)
         const { title, pinned, color } = data;
 
-        const query = `UPDATE to_do_lists SET title = $1, pinned = $2, color = $3 WHERE id = $4 RETURNING *`;
+        const query = `UPDATE to_do_lists SET title = $1, pinned = $2, color = $3 WHERE id = $4 AND user_id = $5 RETURNING *`;
 
-        const values = [title, pinned || false, color || 'grey', id];
+        const values = [title, pinned || false, color || 'grey', id, userId];
 
         const result = await pool.query(query, values);
         return result.rows[0];
     }
 
-    async delete(id) {
+    async delete(id, userId) {
         const client = await pool.connect();
     
         try {
             // J'utilise BEGIN, COMMIT, ROLLBACK pour éviter que tout soit perdu en cas de problème.
             await client.query('BEGIN');
     
-            const result = await client.query('DELETE FROM to_do_lists WHERE id = $1 RETURNING *', [id]);
+            const result = await client.query('DELETE FROM to_do_lists WHERE id = $1 AND user_id = $2 RETURNING *', [id, userId]);
     
             await client.query('COMMIT');
     
@@ -59,19 +60,19 @@ class Method {
         }
     }
 
-    async countTasksInList(listid) {
-        const query = `SELECT COUNT(id) FROM tasks WHERE to_do_list_id = $1`;
+    async countTasksInList(listid, userId) {
+        const query = `SELECT COUNT(tasks.id) FROM tasks INNER JOIN to_do_lists ON tasks.to_do_list_id = to_do_lists.id WHERE to_do_lists.id = $1 AND to_do_lists.user_id = $2`;
 
-        const values = [listid];
+        const values = [listid, userId];
 
         const result = await pool.query(query, values);
         return result.rows[0].count;
     }
 
-    async countDoneTasksInList(listid) {
-        const query = `SELECT COUNT(id) FROM tasks WHERE to_do_list_id = $1 AND done = $2`;
+    async countDoneTasksInList(listid, userId) {
+        const query = `SELECT COUNT(tasks.id) FROM tasks INNER JOIN to_do_lists ON tasks.to_do_list_id = to_do_lists.id WHERE to_do_lists.id = $1 AND to_do_lists.user_id = $2 AND tasks.done = $3`;
 
-        const values = [listid, true];
+        const values = [listid, userId, true];
 
         const result = await pool.query(query, values);
         return result.rows[0].count;
